15.add-policy-post

+++ app/Http/Controllers/Api/PostsController.php
@@ -33,6 +33,9 @@ class PostsController extends Controller
     public function update(postRequestStore $request)
     {
         $post=Post::find($request->id);
+
+        $this->authorize('update', $post);
+
         $post->title=$request->title;
         $post->body=$request->body;
         if($post->save()){
@@ -42,8 +45,10 @@ class PostsController extends Controller
 
 
     }
-    public function delete(Post $post)
+    public function delete(Request $request,Post $post)
     {
+        $this->authorize('update', $post);
+        
         if($post->delete()){
             $posts=Post::orderBy('id','DESC')->paginate(5);
             return PostResource::collection($posts);
